penalty optimization and toast for new best time

+++ components/timerArea/timer.tsx
@@ -2,12 +2,15 @@ import { useScramble } from '@/context/ScrambleContext'
 import { deleteSolve, getSolves, saveSolve, updateSolveTime } from '@/database/database'
 import { useSolves } from '@/hooks/useSolves'
 import React, { useEffect, useState } from 'react'
-import { Pressable, StyleSheet, Text, View } from 'react-native'
+import { Alert, Pressable, StyleSheet, Text, View } from 'react-native'
+import Toast from 'react-native-toast-message'
 
 type Props = {
   fullscreen: boolean
   setFullscreen: React.Dispatch<React.SetStateAction<boolean>>
 }
+
+type Penalty = 'none' | '+2' | 'DNF'
 
 const Timer = ({fullscreen, setFullscreen}: Props) => {
   const [time, setTime] = useState('0.00')
@@ -17,8 +20,8 @@ const Timer = ({fullscreen, setFullscreen}: Props) => {
   const [startTime, setStartTime] = useState(0)
   const {scramble, nextScramble} = useScramble()
   const [lastSolveId, setLastSolveId] = useState<number | null>(null)
-  const [plusTwoActive, setPlusTwoActive] = useState(false);
-  const { solves, bestTime, refreshSolves, formatTime } = useSolves()
+  const [penalty, setPenalty] = useState<Penalty>('none')
+  const { solves, bestTime, refreshSolves, formatTime, parseTime } = useSolves()
 
   useEffect(() => {
     let interval: ReturnType<typeof setInterval>
@@ -45,18 +48,29 @@ const Timer = ({fullscreen, setFullscreen}: Props) => {
   const stopTimer = () => {
     if (running) {
       setRunning(false)
-      saveSolve(time,scramble)
-      setTime(time)
+      saveSolve(time, scramble)
+
       const newBest = refreshSolves()
-      setPlusTwoActive(false); // reset +2
       const latest = getSolves()[0]
       setLastSolveId(latest?.id ?? null)
+
+      if (newBest !== null && latest) {
+        const latestTimeSec = parseTime(latest.time)
+        if (latestTimeSec === newBest) {
+          Toast.show({
+            type: 'success',
+            text1: 'New Best Time!',
+            text2: latest.time,
+            position: 'bottom',
+            bottomOffset: 200,
+          })
+        }
+      }
+
+      setTime(time)
+      setPenalty('none')
       setFullscreen(false)
       nextScramble()
-      console.log(newBest)
-      console.log(
-        getSolves().map(s => ({ id: s.id, time: s.time }))
-      )
     }
     setHolding(true)
   }
@@ -72,32 +86,78 @@ const Timer = ({fullscreen, setFullscreen}: Props) => {
   }
 
   const deleteTime = () => {
-    if (!lastSolveId) return
-    deleteSolve(lastSolveId)
-    setLastSolveId(null)
-    refreshSolves()
+  if (!lastSolveId) return
+
+  Alert.alert(
+    'Delete Solve',
+    'Are you sure you want to delete this solve?',
+    [
+      {
+        text: 'Cancel',
+        style: 'cancel',
+      },
+      {
+        text: 'Yes',
+        onPress: () => {
+          deleteSolve(lastSolveId)
+          setLastSolveId(null)
+          refreshSolves()
+          setPenalty('none')
+          setTime('0.00')
+        },
+        style: 'destructive',
+      },
+    ],
+    { cancelable: true }
+  )
+
   }
 
   const addTwo = () => {
     if (!lastSolveId) return
+
     const last = solves.find(s => s.id === lastSolveId)
     if (!last) return
-    let newTimeSec = last.timeSec
-    if (!plusTwoActive) {
-      newTimeSec += 2
+
+    let newPenalty: Penalty
+
+    if (penalty === '+2') {
+      newPenalty = 'none'
     } else {
-      newTimeSec -= 2
+      newPenalty = '+2'
     }
-    const newTimeStr = formatTime(newTimeSec)
-    updateSolveTime(last.id, newTimeStr)
-    setPlusTwoActive(!plusTwoActive)
-    refreshSolves()
+    applyPenalty(last, newPenalty)
   }
 
-  const didNotFinish = () => {
+  const toggleDNF = () => {
     if (!lastSolveId) return
-    const last = solves[0]
-    updateSolveTime(last.id, 'DNF')
+
+    const last = solves.find(s => s.id === lastSolveId)
+    if (!last) return
+
+    let newPenalty: Penalty
+
+    if (penalty === 'DNF') {
+      newPenalty = 'none'
+    } else {
+      newPenalty = 'DNF'
+    }
+    applyPenalty(last, newPenalty)
+  }
+
+  const applyPenalty = (solve: any, newPenalty: Penalty) => {
+    let newTimeStr: string
+
+    if (newPenalty === 'DNF') {
+      newTimeStr = 'DNF'
+    } else if (newPenalty === '+2') {
+      newTimeStr = formatTime(solve.timeSec + 2)
+    } else {
+      newTimeStr = formatTime(solve.timeSec)
+    }
+
+    updateSolveTime(solve.id, newTimeStr)
+    setPenalty(newPenalty)
     refreshSolves()
   }
 
@@ -119,15 +179,21 @@ const Timer = ({fullscreen, setFullscreen}: Props) => {
         <View style={styles.buttonWrapper}>
           <Pressable onPress={addTwo} style={({ pressed }) => [
             styles.buttonArea,
-            { opacity: pressed ? 0.5 : 1 }
+            { 
+              opacity: pressed ? 0.5 : 1 ,
+              backgroundColor: penalty === '+2' ? '#309164' : 'transparent'
+            }
           ]}>
             <Text style={styles.buttonText}>
               +2
             </Text>
           </Pressable>
-          <Pressable onPress={didNotFinish} style={({ pressed }) => [
+          <Pressable onPress={toggleDNF} style={({ pressed }) => [
             styles.buttonArea,
-            { opacity: pressed ? 0.5 : 1 }
+            { 
+              opacity: pressed ? 0.5 : 1,
+              backgroundColor: penalty === 'DNF' ? '#913030' : 'transparent'
+             }
           ]}>
             <Text style={styles.buttonText}>
               DNF
